+++ patches/src/main/kotlin/app/morphe/patches/tiktok/interaction/sharesheet/ShareSheetToolsPatch.kt
@@ -27,6 +27,7 @@ val shareSheetToolsPatch = bytecodePatch(
     compatibleWith(*AppCompatibilities.tiktok4623())
 
     execute {
+        hookShareModel()
         SettingsStatusLoadFingerprint.method.addInstruction(
             0,
             "invoke-static {}, " +